# services/api/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ObservaX Query API",
    "version": "1.0.0"
  },
  "paths": {
    "/traces/latest": {
      "get": {
        "summary": "Returns recent traces",
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Recent traces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "traces": { "type": "array", "items": { "type": "object" } },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "offset": { "type": "integer" },
                        "count": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": { "description": "Error response" }
        }
      }
    },
    "/traces/{id}": {
      "get": {
        "summary": "Returns trace waterfall",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Trace waterfall" }
        }
      }
    },
    "/logs/search": {
      "get": {
        "summary": "Basic log search",
        "parameters": [
          { "name": "q", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Log search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "logs": { "type": "array", "items": { "type": "object" } },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "offset": { "type": "integer" },
                        "count": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": { "description": "Error response" }
        }
      }
    },
    "/metrics/query": {
      "post": {
        "summary": "Returns metric series",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "metric": { "type": "string" },
                  "start": { "type": "string" },
                  "end": { "type": "string" },
                  "filters": { "type": "object" },
                  "limit": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Metric series",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "metrics": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          },
          "500": { "description": "Error response" }
        }
      }
    },
    "/services": {
      "get": {
        "summary": "Lists detected services",
        "responses": {
          "200": {
            "description": "Service list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "services": { "type": "array", "items": { "type": "string" } }
                  }
                }
              }
            }
          },
          "500": { "description": "Error response" }
        }
      }
    }
  }
}